+++ src/main/java/com/tsystems/ecare/app/dao/PlanDao.java
@@ -5,5 +5,5 @@ import com.tsystems.ecare.app.model.Plan;
 /**
  * @author Andrei Makarevich
  */
-public interface PlanDao extends GenericDao<Plan, Integer> {
+public interface PlanDao extends GenericDao<Plan, Long> {
 }
